Replace u64 with usize for index store

+++ crates/vizia_core/src/binding/lens.rs
@@ -344,7 +344,7 @@ where
     }
 
     fn id(&self) -> StoreId {
-        StoreId::Index(TypeId::of::<Self>(), self.index as u64)
+        StoreId::Index(TypeId::of::<Self>(), self.index)
     }
 }
